Fix NPE and add test
